Create generic class for testing the editor.
