+++ osu.Game.Rulesets.Karaoke.Tests/Screens/Edit/Beatmap/TestSceneKaraokeBeatmapEditor.cs
@@ -1,48 +1,10 @@
 // Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using osu.Framework.Allocation;
-using osu.Framework.Graphics;
-using osu.Framework.Graphics.Containers;
-using osu.Game.Overlays;
-using osu.Game.Rulesets.Edit;
-using osu.Game.Rulesets.Karaoke.Beatmaps;
 using osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps;
-using osu.Game.Rulesets.Karaoke.Tests.Beatmaps;
-using osu.Game.Screens.Edit;
 
 namespace osu.Game.Rulesets.Karaoke.Tests.Screens.Edit.Beatmap;
 
-public partial class TestSceneKaraokeBeatmapEditor : ScreenTestScene<KaraokeBeatmapEditor>
+public partial class TestSceneKaraokeBeatmapEditor : GenericEditorTestScene<KaraokeBeatmapEditor, KaraokeBeatmapEditorScreenMode>
 {
-    [Cached(typeof(EditorBeatmap))]
-    [Cached(typeof(IBeatSnapProvider))]
-    private readonly EditorBeatmap editorBeatmap;
-
-    protected override Container<Drawable> Content { get; } = new Container { RelativeSizeAxes = Axes.Both };
-
-    protected override KaraokeBeatmapEditor CreateScreen() => new();
-
-    private DialogOverlay dialogOverlay = null!;
-
-    public TestSceneKaraokeBeatmapEditor()
-    {
-        var beatmap = new TestKaraokeBeatmap(new KaraokeRuleset().RulesetInfo);
-        var karaokeBeatmap = new KaraokeBeatmapConverter(beatmap, new KaraokeRuleset()).Convert() as KaraokeBeatmap;
-        editorBeatmap = new EditorBeatmap(karaokeBeatmap);
-    }
-
-    [BackgroundDependencyLoader]
-    private void load()
-    {
-        Beatmap.Value = CreateWorkingBeatmap(editorBeatmap.PlayableBeatmap);
-
-        base.Content.AddRange(new Drawable[]
-        {
-            Content,
-            dialogOverlay = new DialogOverlay(),
-        });
-
-        Dependencies.CacheAs<IDialogOverlay>(dialogOverlay);
-    }
 }
